+++ src/background/state/application/icons.rs
@@ -8,7 +8,7 @@ use itertools::Itertools;
 use seelen_core::{
     handlers::SeelenEvent,
     resource::ResourceText,
-    state::{Icon, IconPack},
+    state::{AppIconPackEntry, CustomIconPackEntry, FileIconPackEntry, Icon, IconPack},
 };
 use tauri::Emitter;
 
@@ -38,55 +38,98 @@ impl IconPacksManager {
         self.0.get_mut("system").unwrap()
     }
 
-    pub fn add_system_app_icon(&mut self, key: &str, icon: Icon) {
+    pub fn add_system_app_icon(&mut self, umid: Option<&str>, path: Option<&Path>, icon: Icon) {
+        if umid.is_none() && path.is_none() {
+            return;
+        }
         let system_pack = self.get_system_mut();
-        system_pack.apps.insert(key.to_string(), icon);
+        system_pack.app_entries.push(AppIconPackEntry {
+            umid: umid.map(|s| s.to_string()),
+            path: path.map(|p| p.to_path_buf()),
+            redirect: None,
+            icon: Some(icon),
+        });
+    }
+
+    pub fn add_system_icon_redirect(
+        &mut self,
+        umid: Option<String>,
+        origin: &Path,
+        redirect: &Path,
+    ) {
+        let system_pack = self.get_system_mut();
+        system_pack.app_entries.push(AppIconPackEntry {
+            umid,
+            path: Some(origin.to_path_buf()),
+            redirect: Some(redirect.to_path_buf()),
+            icon: None,
+        });
     }
 
     pub fn add_system_file_icon(&mut self, origin_extension: &str, icon: Icon) {
         let system_pack = self.get_system_mut();
-        system_pack.files.insert(origin_extension.to_string(), icon);
+        system_pack.file_entries.push(FileIconPackEntry {
+            extension: origin_extension.to_string(),
+            icon,
+        });
     }
 
     fn icon_exists(&self, icon: &Icon) -> bool {
-        let root = SEELEN_COMMON
-            .user_icons_path()
-            .join(&self.get_system().metadata.filename);
-        match icon {
-            Icon::Static(path) => root.join(path).exists(),
-            Icon::Dynamic {
-                light,
-                dark,
-                mask: _,
-            } => root.join(light).exists() && root.join(dark).exists(),
-        }
+        let root = SEELEN_COMMON.user_icons_path().join("system");
+        icon.base.as_ref().is_some_and(|p| root.join(p).exists())
+            || (icon.light.as_ref().is_some_and(|p| root.join(p).exists())
+                && icon.dark.as_ref().is_some_and(|p| root.join(p).exists()))
     }
 
     /// Get icon pack by app user model id, filename or path
-    pub fn get_app_icon(&self, key: &str) -> Option<&Icon> {
-        let filename = PathBuf::from(key)
-            .file_name()
-            .map(|p| p.to_string_lossy().to_string());
-
+    pub fn has_app_icon(&self, umid: Option<&str>, path: Option<&Path>) -> bool {
         let icon_pack = self.get_system();
-        let maybe_icon = icon_pack.apps.get(key).or_else(|| match filename.as_ref() {
-            Some(filename) => icon_pack.apps.get(filename),
-            None => None,
-        });
-        if let Some(icon) = maybe_icon {
-            if self.icon_exists(icon) {
-                return Some(icon);
+        let lower_path = path.map(|p| p.to_string_lossy().to_lowercase());
+
+        for entry in &icon_pack.app_entries {
+            let mut found = None;
+            if let (Some(entry_umid), Some(umid)) = (&entry.umid, umid) {
+                if entry_umid == umid {
+                    found = Some(entry);
+                }
             }
+
+            if found.is_none()
+                && entry
+                    .path
+                    .as_ref()
+                    .map(|p| p.to_string_lossy().to_lowercase())
+                    == lower_path
+            {
+                found = Some(entry);
+            }
+
+            if let Some(entry) = found {
+                if entry.redirect.is_some() {
+                    return true;
+                }
+
+                if let Some(icon) = &entry.icon {
+                    if self.icon_exists(icon) {
+                        return true;
+                    }
+                }
+            };
         }
-        None
+
+        false
     }
 
     pub fn get_file_icon(&self, path: &Path) -> Option<&Icon> {
         let extension = path.extension()?.to_string_lossy().to_lowercase();
         let icon_pack = self.get_system();
-        if let Some(icon) = icon_pack.files.get(extension.as_str()) {
-            if self.icon_exists(icon) {
-                return Some(icon);
+        if let Some(entry) = icon_pack
+            .file_entries
+            .iter()
+            .find(|e| e.extension.to_lowercase() == extension)
+        {
+            if self.icon_exists(&entry.icon) {
+                return Some(&entry.icon);
             }
         }
         None
@@ -94,9 +137,9 @@ impl IconPacksManager {
 
     pub fn clear_system_icons(&mut self) -> Result<()> {
         let system_pack = self.get_system_mut();
-        system_pack.apps.clear();
-        system_pack.files.clear();
-        system_pack.specific.clear();
+        system_pack.app_entries.clear();
+        system_pack.file_entries.clear();
+        system_pack.custom_entries.clear();
         let meta = std::ffi::OsStr::new("metadata.yml");
         for entry in std::fs::read_dir(SEELEN_COMMON.user_icons_path().join("system"))?.flatten() {
             if entry.file_type()?.is_dir() {
@@ -135,6 +178,7 @@ impl IconPacksManager {
         let folder_path = SEELEN_COMMON
             .user_icons_path()
             .join("system/folder-icon.svg");
+        let url_path = SEELEN_COMMON.user_icons_path().join("system/url.png");
 
         if !missing_path.exists() || initial {
             std::fs::copy(
@@ -163,24 +207,49 @@ impl IconPacksManager {
             )?;
         }
 
-        system_pack.missing = Some(Icon::Static(PathBuf::from("missing-icon.png")));
-        system_pack.specific.insert(
-            "@seelen/weg::start-menu".to_string(),
-            Icon::Dynamic {
-                light: PathBuf::from("start-menu-icon.svg"),
-                dark: PathBuf::from("start-menu-icon.svg"),
-                mask: Some(PathBuf::from("start-menu-icon.svg")),
-            },
-        );
-        system_pack.specific.insert(
-            "@seelen/weg::folder".to_string(),
-            Icon::Dynamic {
-                light: PathBuf::from("folder-icon.svg"),
-                dark: PathBuf::from("folder-icon.svg"),
-                mask: Some(PathBuf::from("folder-icon.svg")),
-            },
-        );
+        if !url_path.exists() || initial {
+            std::fs::copy(
+                SEELEN_COMMON
+                    .app_resource_dir()
+                    .join("static/icons/url.png"),
+                url_path,
+            )?;
+        }
 
+        system_pack.missing = Some(Icon {
+            base: Some(PathBuf::from("missing-icon.png")),
+            ..Default::default()
+        });
+
+        if !system_pack
+            .file_entries
+            .iter()
+            .any(|e| e.extension == "url")
+        {
+            system_pack.file_entries.push(FileIconPackEntry {
+                extension: "url".to_string(),
+                icon: Icon {
+                    base: Some(PathBuf::from("url.png")),
+                    ..Default::default()
+                },
+            });
+        }
+
+        let mut add_if_none = |key: &str, value: &str| {
+            if !system_pack.custom_entries.iter().any(|e| e.key == key) {
+                system_pack.custom_entries.push(CustomIconPackEntry {
+                    key: key.to_owned(),
+                    icon: Icon {
+                        base: Some(PathBuf::from(value)),
+                        mask: Some(PathBuf::from(value)),
+                        ..Default::default()
+                    },
+                });
+            };
+        };
+
+        add_if_none("@seelen/weg::start-menu", "start-menu-icon.svg");
+        add_if_none("@seelen/weg::folder", "folder-icon.svg");
         Ok(())
     }
 
@@ -208,14 +277,6 @@ impl FullState {
         Ok(())
     }
 
-    fn load_icon_pack_from_dir(dir_path: &Path) -> Result<IconPack> {
-        let file = dir_path.join("metadata.yml");
-        if !file.exists() {
-            return Err("metadata.yml not found".into());
-        }
-        Ok(serde_yaml::from_str(&std::fs::read_to_string(&file)?)?)
-    }
-
     pub(super) fn load_icons_packs(&mut self, initial: bool) -> Result<()> {
         let entries = std::fs::read_dir(SEELEN_COMMON.user_icons_path())?;
         let mut icon_packs_manager = trace_lock!(self.icon_packs);
@@ -223,18 +284,13 @@ impl FullState {
 
         for entry in entries.flatten() {
             let path = entry.path();
-            if !path.is_dir() || !path.join("metadata.yml").exists() {
-                continue;
-            }
-
-            let mut icon_pack = match Self::load_icon_pack_from_dir(&path) {
+            let mut icon_pack = match IconPack::load(&path) {
                 Ok(icon_pack) => icon_pack,
                 Err(err) => {
                     log::error!("Failed to load icon pack ({path:?}): {err:?}");
                     continue;
                 }
             };
-
             icon_pack.metadata.bundled = entry.file_name() == "system";
             icon_pack.metadata.filename = entry.file_name().to_string_lossy().to_string();
             icon_packs_manager